+++ src/components/searchBar.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Box } from "@chakra-ui/layout";
-import { DefaultInput } from "./Inputs";
-import { PrimaryButton } from "./buttons/Button";
+import { DefaultInput } from "./inputs";
+import { PrimaryButton } from "./buttons/button";
 import { MdCall, MdPerson, MdGroups } from "react-icons/md";
-import { IconSelect } from "./Select";
+import { IconSelect } from "./select";
 import { InputGroup, InputLeftElement } from "@chakra-ui/input";
 
 export const SearchBar = () => {
