+++ src/data/mockdata.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { AddCustomerDrawer } from "../components/drawer/CustomerDrawer";
-import { AddGroupDrawer } from "../components/drawer/GroupDrawer";
-import { GroupModalBtn } from "../components/modal/Customermodal";
+import { AddCustomerDrawer } from "../components/drawer/customerdrawer";
+import { AddGroupDrawer } from "../components/drawer/groupdrawer";
 
 export const customerMockData = {
     OptionHeaderData: {
