+++ src/page/group.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Box } from "@chakra-ui/layout";
-import { DefaultHeader } from "../components/Headers";
+import { DefaultHeader } from "../components/headers";
 import { groupMockData } from "../data/mockdata";
-import { GroupTable } from "../components/table/GroupTable";
-import { Pagination } from "../components/Pagination";
+import { GroupTable } from "../components/table/grouptable";
+import { Pagination } from "../components/pagination";
 
 const GroupListPage = () => {
     return (
